remove unused serialno field handling
The serialno field was already explicitly forbidden for the stable and
beta channels and we've never used this with internal release channels.

We don't have a use case for this field so it should just be rejected
outright as we're doing in the recovery implementation.

This code wasn't updated to use READ_PRIVILEGED_PHONE_STATE instead of
READ_PHONE_STATE so this code has already rejected any update package
with this field set since Android 10 anyway.