use AndroidX Preference Library
The previously used Theme.DeviceDefault.Settings can no longer be
applied. We need to use a Theme.AppCompat theme (or descendant) with
the new activity.

The new theme is "hardcoded." It's mirroring the Settings theme used in
GrapheneOS. If we change the Settings theme, the Seamless Update Client
theme must also be changed to continue to provide a consistent user
interface.