Update drawables with new Material Symbols
The file names may seem overly verbose at first, but they contain all
the arguments used for the variable font capability. The outlined style
was used as a basis, and the defaults were kept.

The foreground launcher icon uses the rounded style as a basis and is
filled for consistency with the other app launcher icons from
GrapheneOS.

Remove no longer needed resources.