ignore constraints for user-initiated checks
We intentionally pass the network object even if it's null to trigger an
exception in the Service, allowing us to utilize the user-facing
feedback it provides.

If the network _becomes_ unavailable while the network object is **not**
null, it triggers a SocketException. However, this will be caught by the
Service, and the user will be notified as well. Creating a clearer error
message for this situation would be crafty, so we rely on the exception
message alone.

```
failed to download and install update
java.net.SocketException: Binding socket to network 100 failed: EPERM (Operation not permitted)
```

We'll see whether further improvements are necessary for a better user
experience in this case.

Closes: https://github.com/GrapheneOS/platform_packages_apps_Updater/issues/81